+++ airflow/providers/apache/druid/hooks/druid.py
@@ -24,7 +24,7 @@ from pydruid.db import connect
 
 from airflow.exceptions import AirflowException
 from airflow.hooks.base import BaseHook
-from airflow.hooks.dbapi_hook import DbApiHook
+from airflow.hooks.dbapi import DbApiHook
 
 
 class DruidHook(BaseHook):
