+++ airflow/providers/google/cloud/hooks/bigquery.py
@@ -53,7 +53,7 @@ from pandas_gbq.gbq import (
 )
 
 from airflow.exceptions import AirflowException
-from airflow.hooks.dbapi_hook import DbApiHook
+from airflow.hooks.dbapi import DbApiHook
 from airflow.providers.google.common.hooks.base_google import GoogleBaseHook
 from airflow.utils.helpers import convert_camel_to_snake
 from airflow.utils.log.logging_mixin import LoggingMixin
