+++ airflow/providers/microsoft/mssql/hooks/mssql.py
@@ -19,7 +19,7 @@
 
 import pymssql
 
-from airflow.hooks.dbapi_hook import DbApiHook
+from airflow.hooks.dbapi import DbApiHook
 
 
 class MsSqlHook(DbApiHook):
