+++ airflow/providers/vertica/hooks/vertica.py
@@ -19,7 +19,7 @@
 
 from vertica_python import connect
 
-from airflow.hooks.dbapi_hook import DbApiHook
+from airflow.hooks.dbapi import DbApiHook
 
 
 class VerticaHook(DbApiHook):
