+++ tests/providers/presto/hooks/test_presto.py
@@ -163,7 +163,7 @@ class TestPrestoHook(unittest.TestCase):
 
         self.db_hook = UnitTestPrestoHook()
 
-    @patch('airflow.hooks.dbapi_hook.DbApiHook.insert_rows')
+    @patch('airflow.hooks.dbapi.DbApiHook.insert_rows')
     def test_insert_rows(self, mock_insert_rows):
         table = "table"
         rows = [("hello",), ("world",)]
